hack/services/scripts_manager/requirements.txt
django
djangorestframework
drf-yasg
requests
